Делаю рейтинг

+++ app/src/main/java/com/example/projectexcursions/net/ApiService.kt
@@ -91,6 +91,6 @@ interface ApiService {
     @POST("excursion/rating")
     suspend fun uploadRating(
         @Query("excursionId") excursionId: Long,
-        @Query("ratingValue") ratingValue: BigDecimal
+        @Query("ratingValue") ratingValue: Float
     ) : RatingResponse
 }
